+++ src/main/java/pl/put/poznan/transformer/logic/transforms/CapitalizeTextTransformerDecorator.java
@@ -1,4 +1,6 @@
-package pl.put.poznan.transformer.logic;
+package pl.put.poznan.transformer.logic.transforms;
+
+import pl.put.poznan.transformer.logic.Transformer;
 
 /**
  * The CapitalizeTextTransformerDecorator class is a decorator that capitalizes the first letter of the text.
