+++ src/main/java/pl/put/poznan/transformer/logic/transforms/TextTransformerDecorator.java
@@ -1,4 +1,6 @@
-package pl.put.poznan.transformer.logic;
+package pl.put.poznan.transformer.logic.transforms;
+
+import pl.put.poznan.transformer.logic.Transformer;
 
 /**
  * The TextTransformerDecorator class is a base class for decorators implementing the Transformer interface.
